+++ airbnbapi/src/main/java/com/project/service/RoomServiceImpl.java
@@ -3,12 +3,15 @@ package com.project.service;
 import com.project.dto.RoomDto;
 import com.project.entity.Hotel;
 import com.project.entity.Room;
+import com.project.entity.User;
 import com.project.exception.ResourceNotFoundException;
+import com.project.exception.UnAuthorizedAccessException;
 import com.project.repository.HotelRepository;
 import com.project.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,6 +34,11 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+            throw new UnAuthorizedAccessException("User does not own this hotel");
+
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -49,6 +57,11 @@ public class RoomServiceImpl implements RoomService{
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner()))
+            throw new UnAuthorizedAccessException("User does not own this hotel");
+
+
         return hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper.map(element, RoomDto.class))
@@ -61,6 +74,11 @@ public class RoomServiceImpl implements RoomService{
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner()))
+            throw new UnAuthorizedAccessException("User does not own this hotel");
+
         return modelMapper.map(room, RoomDto.class);
     }
 
@@ -71,6 +89,11 @@ public class RoomServiceImpl implements RoomService{
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner()))
+            throw new UnAuthorizedAccessException("User does not own this hotel");
+
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
     }
